Remove EBI reference
I modified the sentence regarding life science repositories, since it seems a bit double to have both ELIXIR and EBI reference there (they are basically the same anyway). Also, having only the ELIXIR reference makes it more correct with what is actually listed since it is not a complete overlap.